Detect tag from package script.

+++ package.sh
@@ -1,6 +1,16 @@
 #!/bin/sh -e
 
-( cd src && npm install )
+tag="$(git tag --points-at HEAD)"
+target="deval-$tag.zip"
+
+if [ -z "$tag" ]; then
+	echo >&2 "error: HEAD doesn't point to a tag"
+	exit 1
+fi
+
+( cd src && npm install --silent )
 ln -s src deval
-zip -qr deval.zip deval -x '*.pegjs'
+zip -qr "$target" deval -x '*.pegjs'
 rm deval
+
+echo "built release active: $target"
